Move migrations test to own file
